code stylechanges

+++ src/main/java/edu/bu/met/cs665/Americano.java
@@ -6,7 +6,7 @@ public class Americano extends Coffee{
         super(milk, sugar);
     }
 
-    public void setAddPrice(int addPrice){
+    public void setAddPrice(float addPrice){
         this.addPrice = addPrice;
     }
