+++ src/main/java/edu/bu/met/cs665/BlackTea.java
@@ -7,7 +7,7 @@ public class BlackTea extends Tea {
         super(milk, sugar);
     }
 
-    public void setAddPrice(int addPrice){
+    public void setAddPrice(float addPrice){
         this.addPrice = addPrice;
     }
