+++ src/main/java/edu/bu/met/cs665/YellowTea.java
@@ -5,7 +5,7 @@ public class YellowTea extends Tea {
     public YellowTea(int milk, int sugar) {
         super(milk, sugar);
     }
-    public void setAddPrice(int addPrice){
+    public void setAddPrice(float addPrice){
         this.addPrice = addPrice;
     }
